+++ src/main/java/com/devstack/app/service/impl/CutomerServiceImpl.java
@@ -24,7 +24,9 @@ public class CutomerServiceImpl implements CustomerService {
         customerDao.saveCustomer(customer);
     }
 
-    public Customer findCustomer(int id) {
+    public Customer findCustomer(long id) {
         return customerDao.findCustomer(id);
     }
+
+
 }
